requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
Django==5.2
django-appconf==1.1.0
django-debug-toolbar==5.2.0
django-filter==25.1
django-imagekit==5.0.0
django-jalali==7.4.0
django-redis==5.4.0
django-render-partial==0.4
gunicorn==23.0.0
idna==3.10
iniconfig==2.1.0
jalali_core==1.0.0
jdatetime==5.2.0
kavenegar==1.1.2
packaging==25.0
pilkit==3.0
pillow==11.2.1
pluggy==1.6.0
Pygments==2.19.1
python-decouple==3.8
redis==6.0.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
uuid==1.30
whitenoise==6.9.0
